services/nlp/requirements.txt
 
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
google-generativeai==0.3.2
numpy==1.24.3
pydantic==2.5.0
 